pyedec.py
```python
# pyedec.py

# Import necessary libraries
import os
import shutil
from git import Repo
import config

# Function to create directory structure
def create_structure(base_path, structure):
    for key, value in structure.items():
        path = os.path.join(base_path, key)
        os.makedirs(path, exist_ok=True)
        if isinstance(value, dict):
            create_structure(path, value)

# Function to initialize git repository
def init_git_repo(path):
    if config.VCS.lower() == 'git':
        Repo.init(path)

# Function to install dependencies
def install_dependencies():
    os.system(f'pip install -r {config.DEPENDENCIES_FILE}')

# Function to generate project
def generate_project():
    project_path = os.path.join(config.BASE_DIR, 'project')
    os.makedirs(project_path, exist_ok=True)
    create_structure(project_path, config.DEFAULT_PROJECT_STRUCTURE)
    init_git_repo(project_path)
    install_dependencies()

# Function to run AI optimization
def run_ai_optimization():
    if config.AI_OPTIMIZATION['enabled']:
        # Implement AI optimization algorithm here
        pass

# Function to adapt structure dynamically
def adapt_structure_dynamically():
    if config.DYNAMIC_STRUCTURE_ADAPTATION['enabled']:
        # Implement dynamic structure adaptation here
        pass

# Function to generate automated documentation
def generate_automated_documentation():
    if config.AUTOMATED_DOCUMENTATION['enabled']:
        # Implement automated documentation generation here
        pass

# Main function
def main():
    generate_project()
    run_ai_optimization()
    adapt_structure_dynamically()
    generate_automated_documentation()

if __name__ == "__main__":
    main()
```
