requirements.txt
numpy==1.21.2
scikit-learn==0.24.2
pandas==1.3.2
GitPython==3.1.18
